+++ crates/core/src/lib.rs
@@ -2,8 +2,8 @@ pub mod accessibility;
 pub mod dom;
 pub mod events;
 pub mod layout;
-pub mod navigation_mode;
 pub mod node;
+pub mod platform_state;
 pub mod plugins;
 pub mod render;
 pub mod types;
@@ -13,16 +13,9 @@ pub mod prelude {
     pub use crate::dom::*;
     pub use crate::events::*;
     pub use crate::layout::*;
-    pub use crate::navigation_mode::*;
     pub use crate::node::*;
+    pub use crate::platform_state::*;
     pub use crate::plugins::*;
     pub use crate::render::*;
-
-    pub use crate::types::AccessibilityId;
-    pub use crate::types::EventEmitter;
-    pub use crate::types::EventReceiver;
-    pub use crate::types::EventsQueue;
-    pub use crate::types::NativePlatformReceiver;
-    pub use crate::types::NativePlatformSender;
-    pub use crate::types::PotentialEvents;
+    pub use crate::types::*;
 }
