+++ crates/core/src/types.rs
@@ -1,6 +1,6 @@
 use crate::{
     events::DomEvent,
-    prelude::{EventName, PlatformEvent, PotentialEvent},
+    prelude::{EventName, NativePlatformState, PlatformEvent, PotentialEvent},
 };
 pub use accesskit::NodeId as AccessibilityId;
 use rustc_hash::FxHashMap;
@@ -8,31 +8,7 @@ use smallvec::SmallVec;
 use tokio::sync::mpsc::{UnboundedReceiver, UnboundedSender};
 use tokio::sync::watch;
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Default)]
-pub enum PreferredTheme {
-    #[default]
-    /// Use the light variant.
-    Light,
-
-    /// Use the dark variant.
-    Dark,
-}
-
-impl From<winit::window::Theme> for PreferredTheme {
-    fn from(value: winit::window::Theme) -> Self {
-        match value {
-            winit::window::Theme::Light => Self::Light,
-            winit::window::Theme::Dark => Self::Dark,
-        }
-    }
-}
-
-pub struct NativePlatformState {
-    pub focused_id: AccessibilityId,
-    pub preferred_theme: PreferredTheme,
-}
-
-/// Send focus updates to the platform
+/// Send platform updates from the platform
 pub type NativePlatformSender = watch::Sender<NativePlatformState>;
 
 /// Receive updates by the platform
