+++ crates/hooks/src/use_focus.rs
@@ -2,7 +2,7 @@ use dioxus_core::{use_hook, AttributeValue};
 use dioxus_hooks::{use_context, use_memo};
 use dioxus_signals::{Memo, Readable, Signal, Writable};
 use freya_core::{
-    accessibility::ACCESSIBILITY_ROOT_ID, navigation_mode::NavigationMode, types::AccessibilityId,
+    accessibility::ACCESSIBILITY_ROOT_ID, platform_state::NavigationMode, types::AccessibilityId,
 };
 use freya_elements::events::{keyboard::Code, KeyboardEvent};
 use freya_node_state::CustomAttributeValues;
