+++ crates/testing/src/launch.rs
@@ -4,10 +4,7 @@ use dioxus_core::VirtualDom;
 use dioxus_core_macro::rsx;
 use freya_common::EventMessage;
 use freya_components::NativeContainer;
-use freya_core::{
-    prelude::*,
-    types::{NativePlatformState, PreferredTheme},
-};
+use freya_core::prelude::*;
 use freya_engine::prelude::*;
 use freya_hooks::PlatformInformation;
 use std::sync::{Arc, Mutex};
@@ -36,7 +33,8 @@ pub fn launch_test_with_config(root: AppComponent, config: TestingConfig) -> Tes
     let (platform_event_emitter, platform_event_receiver) = unbounded_channel::<EventMessage>();
     let (platform_sender, platform_receiver) = watch::channel(NativePlatformState {
         focused_id: ACCESSIBILITY_ROOT_ID,
-        preferred_theme: PreferredTheme::Light,
+        preferred_theme: PreferredTheme::default(),
+        navigation_mode: NavigationMode::default(),
     });
     let mut font_collection = FontCollection::new();
     font_collection.set_dynamic_font_manager(FontMgr::default());
@@ -54,7 +52,6 @@ pub fn launch_test_with_config(root: AppComponent, config: TestingConfig) -> Tes
         platform_event_receiver,
         accessibility_manager: AccessibilityManager::new(ACCESSIBILITY_ROOT_ID).wrap(),
         ticker_sender: broadcast::channel(5).0,
-        navigation_state: NavigatorState::new(NavigationMode::NotKeyboard),
         platform_information: Arc::new(Mutex::new(PlatformInformation::new(config.size))),
         cursor_icon: CursorIcon::default(),
         platform_sender,
